langchain/02-financial-news-analyzer/requirements.txt
# LangChain core
langchain==0.1.0
langchain-anthropic==0.1.1
langchain-core==0.1.10

# HTTP requests (optional - for future real news API)
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Type hints
typing-extensions==4.9.0
